+++ clients/node/src/index.ts
@@ -10,7 +10,7 @@
 export { Bubblegum } from "./session.js";
 export type { Channel, LaunchOptions, AttachOptions, MobileAttachOptions, RecoverArgs } from "./session.js";
 
-export { BridgeClient, spawnBridgeTransport } from "./client.js";
+export { BridgeClient, spawnBridgeTransport, resolveBridgeCommand } from "./client.js";
 export type { Transport, SpawnOptions, BridgeClientOptions } from "./client.js";
 
 export { BridgeError } from "./errors.js";